localchat/localchat/requirements.txt
# Requirements for Single Model Database Assistant
langchain==0.2.14
langchain-ollama==0.1.0
langchain-community==0.2.12
mysql-connector-python==8.2.0
tabulate==0.9.0
requests==2.31.0
